v0.11 (#1736)
* add version validation and remove user schema migrations

* comment

* nits

* start drizzle sync schema

* convert kysely to drizzle

* customized migration

* legacy migrations

* cleanup

* fix bug

* fix bug

* add tests

* statement timeout

* fix pglite test setup

* cleanup

* remove migration 0_11

* more exports

* nits

* fix legacy migrations

* add filters

* nits

* start version 2

* migrate status type

* multichain checkpoint

* checkpoints table

* .

* cleanup

* cleanup realtime

* .

* historical multichain ordering

* rebase

* use chain name for status

* fix tests

* fix tests

* fix

* improve crash recovery

* cleanup

* don't error for dev command

* fix seconds

* cleanup

* nits

* cleanup

* undo status updates

* fix client status

* fix client status

* chains

* update examples

* create ponder

* readme

* docs

* status type

* publicClients type

* fix client status type

* nits

* fix chains with no match in viem

* multichain default

* publish command

* create new notificatin channel on publish

* create or replace view

* rename

* fix names

* fix list and prune command

* views schema

* rename

* --views-schema cli flag

* nits

* allow rpc to be undefined

* nits

* docs

* chore: changeset

* cleanup

* fix merge

* fix empty checkpoints

* fix empty checkpoints

* fix empty checkpoints

* flush cache after setup events

---------

+++ packages/core/src/bin/commands/prune.ts
@@ -1,9 +1,11 @@
 import { createBuild } from "@/build/index.js";
 import {
   type PonderApp,
+  VIEWS,
   createDatabase,
-  getPonderMeta,
+  getPonderMetaTable,
 } from "@/database/index.js";
+import { TABLES } from "@/database/index.js";
 import { sqlToReorgTableName } from "@/drizzle/kit/index.js";
 import { createLogger } from "@/internal/logger.js";
 import { MetricsService } from "@/internal/metrics.js";
@@ -12,7 +14,6 @@ import { createShutdown } from "@/internal/shutdown.js";
 import { createTelemetry } from "@/internal/telemetry.js";
 import { count, eq, inArray, sql } from "drizzle-orm";
 import { unionAll } from "drizzle-orm/pg-core";
-import { pgSchema } from "drizzle-orm/pg-core";
 import type { CliOptions } from "../ponder.js";
 import { createExit } from "../utils/exit.js";
 
@@ -58,15 +59,10 @@ export async function prune({ cliOptions }: { cliOptions: CliOptions }) {
   const database = await createDatabase({
     common,
     // Note: `namespace` is not used in this command
-    namespace: "public",
+    namespace: { schema: "public", viewsSchema: undefined },
     preBuild: buildResult.result,
     schemaBuild: emptySchemaBuild,
   });
-
-  const TABLES = pgSchema("information_schema").table("tables", (t) => ({
-    table_name: t.text().notNull(),
-    table_schema: t.text().notNull(),
-  }));
 
   const ponderSchemas = await database.qb.drizzle
     .select({ schema: TABLES.table_schema, tableCount: count() })
@@ -82,14 +78,19 @@ export async function prune({ cliOptions }: { cliOptions: CliOptions }) {
     )
     .groupBy(TABLES.table_schema);
 
+  const ponderViewSchemas = await database.qb.drizzle
+    .select({ schema: VIEWS.table_schema })
+    .from(VIEWS)
+    .where(eq(VIEWS.table_name, "_ponder_meta"));
+
   const queries = ponderSchemas.map((row) =>
     database.qb.drizzle
       .select({
-        value: getPonderMeta(row.schema).value,
+        value: getPonderMetaTable(row.schema).value,
         schema: sql<string>`${row.schema}`.as("schema"),
       })
-      .from(getPonderMeta(row.schema))
-      .where(eq(getPonderMeta(row.schema).key, "app")),
+      .from(getPonderMetaTable(row.schema))
+      .where(eq(getPonderMetaTable(row.schema).key, "app")),
   );
 
   if (queries.length === 0) {
@@ -111,6 +112,7 @@ export async function prune({ cliOptions }: { cliOptions: CliOptions }) {
   }
 
   const tablesToDrop: string[] = [];
+  const viewsToDrop: string[] = [];
   const schemasToDrop: string[] = [];
   const functionsToDrop: string[] = [];
 
@@ -123,24 +125,51 @@ export async function prune({ cliOptions }: { cliOptions: CliOptions }) {
       continue;
     }
 
-    for (const table of value.table_names) {
-      tablesToDrop.push(`"${schema}"."${table}"`);
-      tablesToDrop.push(`"${schema}"."${sqlToReorgTableName(table)}"`);
-      functionsToDrop.push(`"${schema}"."operation_reorg__${table}"`);
-    }
-    tablesToDrop.push(`"${schema}"."_ponder_meta"`);
-    tablesToDrop.push(`"${schema}"."_ponder_status"`);
+    if (ponderViewSchemas.some((vs) => vs.schema === schema)) {
+      for (const table of value.table_names) {
+        viewsToDrop.push(`"${schema}"."${table}"`);
+      }
+      viewsToDrop.push(`"${schema}"."_ponder_meta"`);
+      if (value.version === "2") {
+        viewsToDrop.push(`"${schema}"."_ponder_checkpoint"`);
+      } else {
+        viewsToDrop.push(`"${schema}"."_ponder_status"`);
+      }
 
-    const tableCount = ponderSchemas.find(
-      (s) => s.schema === schema,
-    )!.tableCount;
+      const tableCount = ponderSchemas.find(
+        (s) => s.schema === schema,
+      )!.tableCount;
 
-    if (schema !== "public" && tableCount <= 2 + value.table_names.length * 2) {
-      schemasToDrop.push(schema);
+      if (schema !== "public" && tableCount <= 2 + value.table_names.length) {
+        schemasToDrop.push(schema);
+      }
+    } else {
+      for (const table of value.table_names) {
+        tablesToDrop.push(`"${schema}"."${table}"`);
+        tablesToDrop.push(`"${schema}"."${sqlToReorgTableName(table)}"`);
+        functionsToDrop.push(`"${schema}"."operation_reorg__${table}"`);
+      }
+      tablesToDrop.push(`"${schema}"."_ponder_meta"`);
+      if (value.version === "2") {
+        tablesToDrop.push(`"${schema}"."_ponder_checkpoint"`);
+      } else {
+        tablesToDrop.push(`"${schema}"."_ponder_status"`);
+      }
+
+      const tableCount = ponderSchemas.find(
+        (s) => s.schema === schema,
+      )!.tableCount;
+
+      if (
+        schema !== "public" &&
+        tableCount <= 2 + value.table_names.length * 2
+      ) {
+        schemasToDrop.push(schema);
+      }
     }
   }
 
-  if (tablesToDrop.length === 0) {
+  if (tablesToDrop.length === 0 && viewsToDrop.length === 0) {
     logger.warn({
       service: "prune",
       msg: "No inactive Ponder apps found in this database.",
@@ -149,23 +178,38 @@ export async function prune({ cliOptions }: { cliOptions: CliOptions }) {
     return;
   }
 
-  await database.qb.drizzle.execute(
-    sql.raw(`DROP TABLE IF EXISTS ${tablesToDrop.join(", ")} CASCADE`),
-  );
+  if (tablesToDrop.length > 0) {
+    await database.qb.drizzle.execute(
+      sql.raw(`DROP TABLE IF EXISTS ${tablesToDrop.join(", ")} CASCADE`),
+    );
 
-  logger.warn({
-    service: "prune",
-    msg: `Dropped ${tablesToDrop.length} tables`,
-  });
+    logger.warn({
+      service: "prune",
+      msg: `Dropped ${tablesToDrop.length} tables`,
+    });
+  }
 
-  await database.qb.drizzle.execute(
-    sql.raw(`DROP FUNCTION IF EXISTS ${functionsToDrop.join(", ")} CASCADE`),
-  );
+  if (viewsToDrop.length > 0) {
+    await database.qb.drizzle.execute(
+      sql.raw(`DROP VIEW IF EXISTS ${viewsToDrop.join(", ")} CASCADE`),
+    );
 
-  logger.warn({
-    service: "prune",
-    msg: `Dropped ${functionsToDrop.length} functions`,
-  });
+    logger.warn({
+      service: "prune",
+      msg: `Dropped ${viewsToDrop.length} views`,
+    });
+  }
+
+  if (functionsToDrop.length > 0) {
+    await database.qb.drizzle.execute(
+      sql.raw(`DROP FUNCTION IF EXISTS ${functionsToDrop.join(", ")} CASCADE`),
+    );
+
+    logger.warn({
+      service: "prune",
+      msg: `Dropped ${functionsToDrop.length} functions`,
+    });
+  }
 
   if (schemasToDrop.length > 0) {
     await database.qb.drizzle.execute(
